libhub-publisher/requirements.txt
PyYAML==6.0
cnvrgv2==1.0.13
